+++ brownie/addresses.py
@@ -27,6 +27,7 @@ OGN = '0x8207c1FfC5B6804F6024322CcF34F29c3541Ae26'
 USDT = '0xdAC17F958D2ee523a2206206994597C13D831ec7'
 USDC = '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48'
 DAI = '0x6b175474e89094c44da98b954eedeac495271d0f'
+CVX = '0x4e3FBD56CD56c3e72c1403e103b45Db9da5B9D2B'
 
 # Other
 UNISWAP_V2_ROUTER = '0x7a250d5630b4cf539739df2c5dacb4c659f2488d'
